renamed tests + add vitest instead of jest
