+++ README.md
@@ -3,7 +3,7 @@
 [![pages-build-deployment](https://github.com/BarthPaleologue/planetEngine/actions/workflows/pages/pages-build-deployment/badge.svg)](https://github.com/BarthPaleologue/planetEngine/actions/workflows/pages/pages-build-deployment)
 [![tauri build and release](https://github.com/BarthPaleologue/CosmosJourneyer/actions/workflows/tauri-release.yml/badge.svg)](https://github.com/BarthPaleologue/CosmosJourneyer/actions/workflows/tauri-release.yml)
 [![ESLint Check](https://github.com/BarthPaleologue/CosmosJourneyer/actions/workflows/eslint.yml/badge.svg)](https://github.com/BarthPaleologue/CosmosJourneyer/actions/workflows/eslint.yml)
-[![Jest Coverage](https://github.com/BarthPaleologue/CosmosJourneyer/actions/workflows/tests.yml/badge.svg)](https://github.com/BarthPaleologue/CosmosJourneyer/actions/workflows/tests.yml)
+[![Tests](https://github.com/BarthPaleologue/CosmosJourneyer/actions/workflows/tests.yml/badge.svg)](https://github.com/BarthPaleologue/CosmosJourneyer/actions/workflows/tests.yml)
 [![License](https://img.shields.io/github/license/BarthPaleologue/planetEngine)](./LICENSE.md)
 
 [![Teaser Video](./coverImages/video.png)](https://youtu.be/5pXZqHRShTE)
