+++ package.json
@@ -16,6 +16,7 @@
         "@types/webpack-env": "^1.18.5",
         "@typescript-eslint/eslint-plugin": "^8.18.1",
         "@typescript-eslint/parser": "^8.18.1",
+        "@vitest/coverage-v8": "3.0.2",
         "@webpack-cli/generators": "^3.0.7",
         "cross-env": "^7.0.3",
         "css-loader": "^7.1.2",
@@ -46,6 +47,7 @@
         "typedoc-plugin-missing-exports": "^3.1.0",
         "typescript": "~5.7.2",
         "typescript-plugin-css-modules": "^5.1.0",
+        "vitest": "^3.0.2",
         "webpack": "^5.97.1",
         "webpack-cli": "^5.1.4",
         "webpack-dev-server": "^5.2.0"
@@ -59,8 +61,8 @@
         "serve:prod": "http-server dist -p 8080",
         "format:check": "prettier --config .prettierrc src//**/*.{js,jsx,json,ts} --check",
         "format": "prettier --config .prettierrc src//**/*.{js,json,ts} --write",
-        "test": "cross-env NODE_OPTIONS=--experimental-vm-modules npx jest",
-        "test:coverage": "cross-env NODE_OPTIONS=--experimental-vm-modules npx jest --coverage",
+        "test": "vitest run",
+        "test:coverage": "vitest run --coverage",
         "build:docs": "typedoc --options typedoc.json",
         "serve:docs": "http-server docs -p 8081",
         "lint:check": "eslint src/ts --config eslint.config.mjs",
