+++ tests/neighborStarSystems.spec.ts
@@ -17,9 +17,9 @@
 
 import { getNeighborStarSystemCoordinates } from "../src/ts/utils/getNeighborStarSystems";
 import { Vector3 } from "@babylonjs/core/Maths/math.vector";
-
 import { starSystemCoordinatesEquals } from "../src/ts/utils/coordinates/universeCoordinates";
 import { getStarSystemCoordinatesFromSeed } from "../src/ts/starSystem/systemSeed";
+import { expect, test } from "vitest";
 
 test("getNeighborStarSystemCoordinates", () => {
     const systemCoordinates = getStarSystemCoordinatesFromSeed({
